+++ src/components/CMakeLists.txt
@@ -1,24 +1,15 @@
-function(add_component componentName)
-    add_library(${componentName} ${componentName}/${componentName}.cpp)
-    target_link_libraries(${componentName} Qt5::Widgets Qt5::Core Qt5::SerialBus)
+function(add_component componentName componentModelName)
+	add_library(${componentName} ${componentName}/${componentName}.cpp ${componentName}/${componentModelName}.cpp)
+    target_link_libraries(${componentName} Qt5::Widgets Qt5::Core Qt5::SerialBus nodes cds-common)
     target_include_directories(${componentName} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
 endfunction()
 
-function(add_component_ui componentName)
-    add_library(${componentName} ${componentName}/${componentName}.cpp ${componentName}/${componentName}.ui)
-    target_link_libraries(${componentName} Qt5::Widgets Qt5::Core Qt5::SerialBus cds-common)
+function(add_component_ui componentName componentModelName)
+	add_library(${componentName} ${componentName}/${componentName}.cpp ${componentName}/${componentModelName}.cpp ${componentName}/${componentName}.ui)
+    target_link_libraries(${componentName} Qt5::Widgets Qt5::Core Qt5::SerialBus nodes cds-common)
     target_include_directories(${componentName} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
 endfunction()
 
-function(add_component_node componentName componentWidget)
-    add_library(${componentName} ${componentName}/${componentName}.cpp)
-    target_link_libraries(${componentName} Qt5::Core Qt5::SerialBus nodes cds-common ${componentWidget})
-    target_include_directories(${componentName} INTERFACE ${CMAKE_CURRENT_SOURCE_DIR})
-endfunction()
-
-add_component(candevice)
-add_component_ui(canrawview)
-add_component_ui(canrawsender)
-add_component_node(candevicemodel candevice)
-add_component_node(canrawsendermodel canrawsender)
-add_component_node(canrawviewmodel canrawview)
+add_component(candevice candevicemodel)
+add_component_ui(canrawview canrawviewmodel)
+add_component_ui(canrawsender canrawsendermodel)
