+++ src/components/canrawsender/canrawsendermodel.cpp
