+++ src/components/canrawsender/canrawsendermodel.h
