+++ src/components/canrawview/canrawviewmodel.cpp
