+++ src/components/canrawview/canrawviewmodel.h
