+++ src/gui/CMakeLists.txt
@@ -7,5 +7,5 @@ set(srcs
 )
 
 add_executable(CANdevStudio ${srcs})
-target_link_libraries(CANdevStudio Qt5::Widgets candevice canrawview canrawsender cds-common nodes candevicemodel canrawsendermodel canrawviewmodel)
+target_link_libraries(CANdevStudio Qt5::Widgets candevice canrawview canrawsender cds-common nodes)
 target_compile_definitions(CANdevStudio PRIVATE $<$<CONFIG:Debug>:CDS_DEBUG=true> $<$<NOT:$<CONFIG:Debug>>:CDS_DEBUG=false>)
